% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_history.R
\name{bbs_history}
\alias{bbs_history}
\title{Examine the Number of BBS Sites Surveyed Each Year}
\usage{
bbs_history(type = "plot")
}
\arguments{
\item{type}{Character string specifying the output format: either \code{"table"}
or \code{"plot"}. Default value \code{type = "plot"}}
}
\value{
A \code{tibble} or a \code{ggplot} showing the number of sites surveyed each
year across regions.
}
\description{
This function returns the number of sites surveyed each year in the BBS
Taiwan project. Sites were mapped into five regions: East, West, South,
North, and Mountain (elevation higher than 1,000 m).
}
\examples{
# Return the number of sites in a table
bbs_history(type = "table")

# Return the number of sites in a bar chart
bbs_history(type = "plot")
}
