% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_plotmap.R
\name{bbs_plotmap}
\alias{bbs_plotmap}
\title{Visualize Species Distribution Across All BBS Sites}
\usage{
bbs_plotmap(target_species)
}
\arguments{
\item{target_species}{Character string specifying the scientific name of
the species of interest. It can accept a single character string, such as
\code{target_species = "紅嘴黑鵯"}, or a vector, such as
\code{target_species = c("紅嘴黑鵯", "白耳畫眉")}. The function can accept up to plotting
5 species in one figure. Use \code{NULL} to return a map of site distribution.}
}
\value{
A \code{ggplot} object showing the distribution map.
}
\description{
This function visualizes the sites surveyed for breeding birds in Taiwan,
highlighting the presence and absence of specific species. It is designed
upon the function \link{bbs_fetch} and \link{bbs_translate}.
}
\examples{
# For single species distribution
bbs_plotmap(target_species = "紅嘴黑鵯")

# For multiple species distribution
bbs_plotmap(target_species = c("紅嘴黑鵯", "白耳畫眉"))

# Simply the distribution of the surveyed sites
bbs_plotmap(target_species = NULL)
}
