% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_poptrend.R
\name{bbs_poptrend}
\alias{bbs_poptrend}
\title{Calculate the population trend based on BBS data}
\usage{
bbs_poptrend(data, zone = NULL)
}
\value{
some info here
}
\description{
Calculate the population trend based on BBS data
}
\examples{
bbs_poptrend(data = bbs_fetch(bbs_translate("金背鳩")))
}
