% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_sites.R
\name{bbs_sites}
\alias{bbs_sites}
\title{All BBS site's coordinate}
\usage{
bbs_sites()
}
\value{
a tibble including the coordinates of all the BBS survey sites
}
\description{
All BBS site's coordinate
}
\examples{
bbs_sites()
}
