% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_stat.R
\name{bbs_stat}
\alias{bbs_stat}
\title{Basic stats of BBS Taiwan occurrence data}
\usage{
bbs_stat(data)
}
\arguments{
\item{data}{a list object derived from the bbs_fetch() function}
}
\value{
a tibble containing basic stats for fetched data
}
\description{
Summarizing basic statistics for BBS data, highlighting the trend of specific species.
}
\examples{
bbs_stat(data =
bbs_fetch(target_species =
c("Psilopogon nuchalis", "Pycnonotus taivanus")))

}
