% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_translate.R
\name{bbs_translate}
\alias{bbs_translate}
\title{Translate Bird Species' Chinese Common Name to Scientific Name}
\usage{
bbs_translate(target_species)
}
\arguments{
\item{target_species}{A single character string or a vector of character strings
representing species' names in Chinese.}
}
\value{
A vector of bird species' scientific names. If the input species name is not
included in the bird list of Taiwan, \code{NA} will be returned. Please check for any typos.
}
\description{
This function is intended for use under \link{bbs_fetch} and \link{bbs_plotmap}.
This function helps users find the scientific names of birds from their
Chinese common names for species found in Taiwan.
}
\examples{
# For a single species
bbs_translate("白頭翁")

# For multiple species
bbs_translate(target_species = c("烏頭翁", "白頭翁", "紅嘴黑鵯", "白耳畫眉"))
}
