% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsTaiwan-package.R
\docType{package}
\name{bbsTaiwan-package}
\alias{bbsTaiwan}
\alias{bbsTaiwan-package}
\title{bbsTaiwan: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sunnytseng.github.io/bbsTaiwan/}
  \item \url{https://github.com/SunnyTseng/bbsTaiwan}
  \item Report bugs at \url{https://github.com/SunnyTseng/bbsTaiwan/issues}
}

}
\author{
\strong{Maintainer}: Sunny Tseng \email{sunnyyctseng@gmail.com}

}
\keyword{internal}
