% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_info.R
\docType{data}
\name{bird_info}
\alias{bird_info}
\title{BBS Taiwan bird list}
\format{
A data frame with 909 rows and 4 columns:
\describe{
\item{scientificName}{scientific name}
\item{chinesenName}{all possible Chinese that were used for the species}
\item{englishName}{english name from Taiwan Wild Bird Federation}
\item{scientificName_t}{scientific name from Taiwan Wild Bird Federation}
...
}
}
\source{
\url{https://drive.google.com/drive/folders/1ex6EDkXv82mpEKcPkOYrQJ_anlu3pI1E}
}
\usage{
bird_info
}
\description{
A list of bird species that recorded in BBS Taiwan, including the
scientific name, Chinese name, English name
}
\keyword{datasets}
