% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\docType{data}
\name{event}
\alias{event}
\title{BBS Taiwan raw dataset on GBIF - event}
\format{
A data frame with 423,139 rows and 18 columns:
\describe{
\item{id}{}
\item{eventID}{}
\item{parentEventID}{}
\item{samplingProtocol}{}
\item{sampleSizeValue}{}
\item{sampleSizeUnit}{}
\item{samplingEffort}{}
\item{eventDate}{}
\item{eventTime}{}
\item{locationID}{}
\item{country}{}
\item{countryCode}{}
\item{locality}{}
\item{decimalLatitude}{}
\item{decimalLongitude}{}
\item{geodeticDatum}{}
\item{coordinateUncertaintyInMeters}{}
\item{coordinatePrecision}{}
...
}
}
\source{
\url{https://www.gbif.org/zh-tw/dataset/f170f056-3f8a-4ef3-ac9f-4503cc854ce0}
}
\usage{
event
}
\description{
A dataframe record BBS Taiwan event info. This is a raw dataset, which was
downloaded directly from GBIF without any wrangling or cleaning.
}
\keyword{datasets}
