% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendedmesurementorfact.R
\docType{data}
\name{extendedmeasurementorfact}
\alias{extendedmeasurementorfact}
\title{BBS Taiwan raw dataset on GBIF - extendedmeasurementorfact}
\format{
A data frame with 1,649,589 rows and 7 columns:
\describe{
\item{id}{}
\item{measurementID}{}
\item{measurementType}{}
\item{measurementValue}{}
\item{measurementDeterminedDate}{}
\item{measurementDeterminedBy}{}
\item{measurementMethod}{}
...
}
}
\source{
\url{https://www.gbif.org/zh-tw/dataset/f170f056-3f8a-4ef3-ac9f-4503cc854ce0}
}
\usage{
extendedmeasurementorfact
}
\description{
A dataframe record BBS Taiwan extendedmeasurementorfact, which contains
info for associated occurrence record, such as time, location. This is a
raw dataset, which was downloaded directly from GBIF without any wrangling
or cleaning.
}
\keyword{datasets}
