% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurementorfacts.R
\docType{data}
\name{measurementorfacts}
\alias{measurementorfacts}
\title{BBS Taiwan raw dataset on GBIF - measurementorfacts}
\format{
A data frame with 1,649,589 rows and 7 columns:
\describe{
\item{id}{}
\item{measurementID}{}
\item{measurementType}{}
\item{measurementValue}{}
\item{measurementDeterminedDate}{}
\item{measurementDeterminedBy}{}
\item{measurementMethod}{}
...
}
}
\source{
\url{https://www.gbif.org/zh-tw/dataset/f170f056-3f8a-4ef3-ac9f-4503cc854ce0}
}
\usage{
measurementorfacts
}
\description{
A dataframe record BBS Taiwan measurementorfacts, which contains info for
associated event record, such as time, location. This is a
raw dataset, which was downloaded directly from GBIF without any wrangling
}
\keyword{datasets}
