% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence.R
\docType{data}
\name{occurrence}
\alias{occurrence}
\title{BBS Taiwan raw dataset on GBIF - occurrence}
\format{
A data frame with 385,131 rows and 11 columns:
\describe{
\item{id}{}
\item{basisOfRecord}{}
\item{occurrenceID}{}
\item{recordedBy}{}
\item{individualCount}{}
\item{occurrenceRemarks}{}
\item{eventID}{}
\item{scientificName}{}
\item{family}{}
\item{genus}{}
\item{vernacularName}{}
...
}
}
\source{
\url{https://www.gbif.org/zh-tw/dataset/f170f056-3f8a-4ef3-ac9f-4503cc854ce0}
}
\usage{
occurrence
}
\description{
A dataframe record BBS Taiwan occurrence info. This is a raw dataset,
which was downloaded directly from GBIF without any wrangling or cleaning.
}
\keyword{datasets}
