% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_elev.R
\docType{data}
\name{tw_elev}
\alias{tw_elev}
\title{Elevation raster of Taiwan}
\format{
A data frame with 38,575 rows and 3 columns:
\describe{
\item{x}{scientific name}
\item{y}{all possible Chinese that were used for the species}
\item{G1km_TWD97-121_DTM_ELE}{english name from Taiwan Wild Bird Federation}
...
}
}
\source{
\url{https://github.com/WanJyunChen/Taiwan_environmental_dataset}
}
\usage{
tw_elev
}
\description{
A raster in dataframe xyz format, representing the elevation of Taiwan in
1m by 1m grids. The data were reprojected to WGS84 (EPSG:4326).
}
\keyword{datasets}
