% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_map.R
\docType{data}
\name{tw_map}
\alias{tw_map}
\title{Map of Taiwan}
\format{
A sf object with 1 feature and 67 fields
}
\source{
\url{https://geodata.libraries.mit.edu}
}
\usage{
tw_map
}
\description{
A sf file created by reading in shape file, showing the outline of Taiwan.
The crs is WGS84 (EPSG:4326).
}
\keyword{datasets}
