% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_region.R
\docType{data}
\name{tw_region}
\alias{tw_region}
\title{Map of eco-regions in Taiwan}
\format{
Need more info
}
\source{
\url{https://ndltd.ncl.edu.tw/cgi-bin/gs32/gsweb.cgi/login?o=dnclcdr&s=id=\%22091NTU00312007\%22.&searchmode=basic}
}
\usage{
tw_region
}
\description{
There are three regions: North, West, and East, representing different
ecosystems in Taiwan specifically for avian biodiversity. This map was
developed by Hau-Jie Shiu in 2003.
}
\keyword{datasets}
