% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_fetch.R
\name{bbs_fetch}
\alias{bbs_fetch}
\title{Fetch BBS Taiwan data from GBIF}
\usage{
bbs_fetch(target_species = NULL)
}
\arguments{
\item{target_species}{a character vector for species of interest, return all species if NULL}
}
\value{
a tibble containing species occurrence
}
\description{
Filter and clean BBS Taiwan data given certain year range and target species
}
\examples{
bbs_fetch(target_species = c("Pycnonotus taivanus", "Pycnonotus sinensis"))
}
