% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_history.R
\name{bbs_history}
\alias{bbs_history}
\title{Number of sites surveyed each year}
\usage{
bbs_history()
}
\value{
A table showing number of sites surveyed each year in regions
}
\description{
Number of sites surveyed each year
}
\examples{
bbs_history()
}
