% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_plotmap.R
\name{bbs_plotmap}
\alias{bbs_plotmap}
\title{Visualization of BBS Taiwan detection sites}
\usage{
bbs_plotmap(data)
}
\arguments{
\item{data}{a list object derived from the bbs_fetch() function}
}
\value{
a distribution map
}
\description{
Visualizes sites surveyed for breeding birds in Taiwan, highlighting presence and absence of specific species.
}
\examples{
bbs_plotmap(data =
bbs_fetch(target_species =
c("Psilopogon nuchalis", "Pycnonotus taivanus")))
}
