% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_plotmap.R
\name{bbs_plotmap}
\alias{bbs_plotmap}
\title{Visualize Species Distribution Across All BBS Sites}
\usage{
bbs_plotmap(data)
}
\arguments{
\item{data}{An occurrence dataset derived from the \link{bbs_fetch} function.}
}
\value{
A \code{ggplot} object showing the distribution map.
}
\description{
This function visualizes the sites surveyed for breeding birds in Taiwan,
highlighting the presence and absence of specific species. It is designed
to be used with the \link{bbs_fetch} function as the input argument.
}
\examples{
plot <- c("Psilopogon nuchalis", "Pycnonotus taivanus") |>
  bbs_fetch() |>
  bbs_plotmap()
}
