% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_plotmap.R
\name{bbs_plotmap}
\alias{bbs_plotmap}
\title{Visualize Species Distribution Across All BBS Sites}
\usage{
bbs_plotmap(target_species = NULL)
}
\arguments{
\item{target_species}{Character string specifying the scientific name of
the species of interest. It can accept a single character string, such as
\code{target_species = "Hypsipetes leucocephalus"}, or a vector, such as
\code{target_species = c("Hypsipetes leucocephalus", "Heterophasia auricularis")}.
Leave undefined or use \code{NULL} to return no species. Use the \link{bbs_translate}
function to help find the species' scientific names.}
}
\value{
A \code{ggplot} object showing the distribution map.
}
\description{
This function visualizes the sites surveyed for breeding birds in Taiwan,
highlighting the presence and absence of specific species. It is designed
upon the function \link{bbs_fetch}.
}
\examples{
bbs_plotmap(target_species = c("Pycnonotus taivanus", "Pycnonotus sinensis"))
}
