% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_poptrend.R
\name{bbs_poptrend}
\alias{bbs_poptrend}
\title{Calculate the population trend based on BBS data}
\usage{
bbs_poptrend(data = NULL, prop = 1)
}
\arguments{
\item{data}{a list object derived from the bbs_fetch() function}

\item{prop}{the proportion of observations to fit the population trend model.
The lower value can reduce the}
}
\value{
some info here
}
\description{
Calculate the population trend based on BBS data
}
\examples{
bbs_poptrend()
}
