% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_translate.R
\name{bbs_translate}
\alias{bbs_translate}
\title{Find species' English common name or scientific name}
\usage{
bbs_translate(species)
}
\arguments{
\item{species}{A character of the bird species' name in Chinese}
}
\value{
A character of the bird species' name in English
}
\description{
Find species' English common name or scientific name
}
\examples{
bbs_translate(species = "烏頭翁")
}
