% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_translate.R
\name{bbs_translate}
\alias{bbs_translate}
\title{Find species' scientific name}
\usage{
bbs_translate(species_name)
}
\arguments{
\item{species_name}{a single character or a vector of characters including species' name in Chinese}
}
\value{
a vector of bird species' scientific name
}
\description{
Find species' scientific name
}
\examples{
# For single species
bbs_translate("白頭翁")

# For multiple species
bbs_translate(c("烏頭翁", "白頭翁"))
}
