% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_region.R
\docType{data}
\name{tw_region}
\alias{tw_region}
\title{Map of regions in Taiwan}
\format{
Need more info
}
\source{
\url{https://geodata.mit.edu/catalog/stanford-dz142zj5454}
}
\usage{
tw_region
}
\description{
Need more info
}
\keyword{datasets}
